requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-localflavor==1.4.1
django-suit==0.2.24
django-toolbelt==0.0.1
ez-setup==0.9
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
static3==0.7.0
